hearMeout-main/digital-psychological-intervention-system/ai-services/requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-google-genai

Flask
Flask-Cors
tenacity
google-api-core

-e .